branch:	HACKT-00-01-04-main-00-80-54-channel-01 rev:	HACKT-00-01-04-main-00-80-54-channel-01-03
  Implemented a connection direction checking policy in
	Object/inst/connection_policy.tcc.
	Correctly rejects mis-connections of channels.
  Tested (successfully) new test cases for directional connections.

+++ test/update-unrollstderr.sh
@@ -23,7 +23,16 @@ done
 
 for i
 do
-	cp -f $i.unrolldump-filter $srcdir/$i.unrollstderr
-	echo "Updated $srcdir/$i.unrollstderr"
+	if test -f $i.unrolldump-filter
+	then
+		cp -f $i.unrolldump-filter $srcdir/$i.unrollstderr
+		echo "Updated $srcdir/$i.unrollstderr"
+	elif test -f $i.unrollfaildump
+	then
+		cp -f $i.unrollfaildump $srcdir/$i.unrollstderr
+		echo "Updated $srcdir/$i.unrollstderr"
+	else
+		echo "$i.unrolldump-filter or $i.unrollfaildump required but missing."
+	fi
 done
